successfully setup db migrations

+++ models/school.py
@@ -8,8 +8,8 @@ class School(db.Model):
     # fields of the Student table
     id = db.Column(db.Integer, primary_key=True)
     name = db.Column(db.String(256), nullable=False, default="")
-    cut_off_points = db.Column(db.Integer(100), nullable=False, default="")
-    number_of_students = db.Column(db.Integer(1000), nullable=False, default="")
+    cut_off_points = db.Column(db.Integer, nullable=False, default="")
+    number_of_students = db.Column(db.Integer, nullable=False, default="")
 
     def __init__(self, name, cut_off_points, number_of_students):
         """ initialize with name, cut off points and no. of Students """
